% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposureMatrix.R
\name{exposureMatrix}
\alias{exposureMatrix}
\title{Convertion of the vector of signatures exposures per genome 
into matrix of signatures exposures per genome.}
\usage{
exposureMatrix(exposure_vector)
}
\arguments{
\item{exposure_vector}{(Mandatory) is a list of objects with 
names of the tumor genomes, the objects are composed each by a 
vector of signatures.}
}
\value{
exposure list composed by matrix type objects.
}
\description{
This function converts the list of vectors 
created with the core function of the package decompTumor2Sig, 
into a list of matrices.
}
\details{
The matrices are composed by number of rows equal to 
the number of signatures and one column, referring to the sample
}
\examples{

## Retrieve signatures and tumor genomes
sigfile <- system.file("extdata",
            paste0("signatures.Rdata"),
            package="VisualComparisonMutationsSetTumors")
genomes_file <- system.file("extdata",
                paste0("Nik-Zainal_PMID_22608084-MPF.Rdata"),
                package="VisualComparisonMutationsSetTumors")

## Find exposures vectors for genome
exposure_vector <- exposureVector(sigfile, genomes_file)

## Convert exposure vectors in exposure matrix
exposure_matrix <- exposureMatrix(exposure_vector)

}
\author{
Michela Dell'Alma\cr Maintainer: Michela Dell'Alma
\cr E-Mail: < micheladellalma@gmail.com> or <michela.dellalma@polimi.it>
}