% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposureVector.R
\name{exposureVector}
\alias{exposureVector}
\title{Exposures vector for each tumor genome, given the MPF file 
and the Alexandrov signatures}
\usage{
exposureVector(sigfile, genomes_file)
}
\arguments{
\item{sigfile}{(Mandatory) Can be a file name or an URL for 
download.}

\item{genomes_file}{(Mandatory) The name of the MPF file 
(can be compressed with \code{gzip}).}
}
\value{
A list of signature weight vectors (also called 
'exposures'), one
for each tumor genome. E.g., the first vector element of 
the first list object is the weight/contribution of the 
first signature to the first tumor genome.
}
\description{
Given as input the tumor genomes as MPF file, 
and a set of Alexandrov signatures, thanks to the function 
already defined in decompTumor2Sig package, the function 
reads the two files and compute the signature weight vectors 
(also called 'exposures'), one for each tumor genome.
}
\examples{

## Retrieve signatures and tumor genomes
sigfile <- system.file("extdata",
            paste0("signatures.Rdata"),
            package="VisualComparisonMutationsSetTumors")
genomes_file <- system.file("extdata",
                paste0("Nik-Zainal_PMID_22608084-MPF.Rdata"),
                package="VisualComparisonMutationsSetTumors")
exposure_vector <- exposureVector(sigfile, genomes_file)

}
\author{
Michela Dell'Alma\cr Maintainer: Michela Dell'Alma
\cr E-Mail: < micheladellalma@gmail.com> or <michela.dellalma@polimi.it>
}
