% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMatrix.R
\name{fillMatrix}
\alias{fillMatrix}
\title{Fill the empty matrix with the values of signature for each genome}
\usage{
fillMatrix(matrix_empty, exposure_matrix)
}
\arguments{
\item{matrix_empty}{(Mandatory) A NA fill matrix with num col equal
to lengthexposure_matrix (num of tumor genomes) and num of rows 
equal to the nrow of the exposure_matrix objects (num signatures).}

\item{exposure_matrix}{(Mandatory) A list of matrices
which length is equal to the number of tumor genomes and 
number of rows in the matrix is equal to the number of signatures}
}
\value{
filled matrix with values of frequencies.
}
\description{
This function fills the empty matrix by coulmns taken 
from the list of exposures per genome. It assumes that the empty 
matrix given has a number of rows equal to the number of rows of 
the object in the list, and a number of columns equal to the length 
of the exposure list (number 
tumor samples).
}
\details{
With a for loop the columns are filled
}
\examples{

#' ## Retrieve signatures and tumor genomes
sigfile <- system.file("extdata",
            paste0("signatures.Rdata"),
            package="VisualComparisonMutationsSetTumors")
genomes_file <- system.file("extdata",
                paste0("Nik-Zainal_PMID_22608084-MPF.Rdata"),
                package="VisualComparisonMutationsSetTumors")

## Find exposures vectors for genome
exposure_vector <- exposureVector(sigfile, genomes_file)

## Convert exposure vectors in exposure matrix
exposure_matrix <- exposureMatrix(exposure_vector)

## Create the empty matrix
matrix_empty <- createEmptyMatrix(exposure_matrix)

## Fill the empty matrix 
matrix_filled <- fillMatrix(matrix_empty, exposure_matrix)

}
\author{
Michela Dell'Alma\cr Maintainer: Michela Dell'Alma
\cr E-Mail: < micheladellalma@gmail.com> or <michela.dellalma@polimi.it>
}
