% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatMap.R
\name{heatMap}
\alias{heatMap}
\title{Draw a heatmap.}
\usage{
heatMap(matrix_named)
}
\arguments{
\item{matrix_named}{(Mandatory) A matrix with as rows the signatures 
names (string) and as columns the tumor genomes (string) and as 
numerical values the contributions.}
}
\value{
Heatmap visualization of the signatures contributions for 
each tumor genome.
}
\description{
This function allows a visualization of the signatures 
weights (exposures) per genome. It assumes that the object of the 
function is a matrix with specific column and row names.
}
\details{
False color image with a dendrogram added to the left side 
and to the top, typically, reordering of the rows and columns according 
to some set of values (row or column means) within the restrictions 
imposed by the dendrogram is carried out.
}
\examples{

#' ## Retrieve signatures and tumor genomes
sigfile <- system.file("extdata",
            paste0("signatures.Rdata"),
            package="VisualComparisonMutationsSetTumors")
genomes_file <- system.file("extdata",
                paste0("Nik-Zainal_PMID_22608084-MPF.Rdata"),
                package="VisualComparisonMutationsSetTumors")

## Find exposures vectors for genome
exposure_vector <- exposureVector(sigfile, genomes_file)

## Convert exposure vectors in exposure matrix
exposure_matrix <- exposureMatrix(exposure_vector)

## Create the empty matrix
matrix_empty <- createEmptyMatrix(exposure_matrix)

## Fill the empty matrix 
matrix_filled <- fillMatrix(matrix_empty, exposure_matrix)

## Name the matrix
matrix_named <- giveNamesMatrix(matrix_filled, exposure_vector)

## Create the heatmap
heatMap(matrix_named)

}
\author{
Michela Dell'Alma\cr Maintainer: Michela Dell'Alma
\cr E-Mail: < micheladellalma@gmail.com> or <michela.dellalma@polimi.it>
}
